refactor: merge osdplayer into fullscreen
